requirements.txt
flask
flask-cors
requests
werkzeug
python-dotenv